# tsdb: sort values for Postings only when required