In the head and in v1 postings on disk, it makes no difference whether postings are sorted. Only for v2 does the code step through in order. So, move the sorting to where it is required, and thus skip it entirely in the head.

Label values in on-disk blocks are already sorted, but `slices.Sort` is very fast on already-sorted data so we don't bother checking.

Update: I added another commit to fast-track `label=""`, which otherwise calls `append()` for every label value.
This required another change to special-case `""=""` which is apparently used in some tests to mean "everything".

Benchmarks (test data only has 100 different label values; the improvement should be much better for thousands):
```
name                                                  old time/op    new time/op    delta
RangeQuery/expr=count({__name__!=""}),steps=1-4         13.5ms ± 1%    13.5ms ± 5%    ~     (p=0.421 n=5+5)
RangeQuery/expr=count({__name__!="",l=""}),steps=1-4     807µs ± 3%     781µs ± 2%  -3.22%  (p=0.032 n=5+5)

name                                                  old alloc/op   new alloc/op   delta
RangeQuery/expr=count({__name__!=""}),steps=1-4         1.43MB ± 0%    1.43MB ± 0%    ~     (p=1.000 n=5+5)
RangeQuery/expr=count({__name__!="",l=""}),steps=1-4    33.4kB ± 0%    33.4kB ± 0%    ~     (p=0.971 n=4+4)

name                                                  old allocs/op  new allocs/op  delta
RangeQuery/expr=count({__name__!=""}),steps=1-4          14.6k ± 0%     14.6k ± 0%    ~     (p=0.571 n=4+5)
RangeQuery/expr=count({__name__!="",l=""}),steps=1-4       432 ± 0%       432 ± 0%    ~     (all equal)
```

Motivation is this profile, where this particular process spends 2.5% of all CPU time in sorting:

![image](https://user-images.githubusercontent.com/8125524/210618570-2770ea59-d5e6-4964-8145-fcf420cc80d5.png)